Use SPI to introduce rule based RQL
